LPS-130175 Move EditableProcessorContext component
